markitwrite/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markitwrite"
version = "0.1.0"
description = "AI virtual clipboard - paste images into any document format (DOCX, PPTX, Markdown)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
all = [
    "python-docx>=1.0.0",
    "python-pptx>=0.6.21",
    "Pillow>=10.0.0",
]
docx = ["python-docx>=1.0.0"]
pptx = ["python-pptx>=0.6.21"]
dev = [
    "pytest>=7.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.21",
    "Pillow>=10.0.0",
]

[project.scripts]
markitwrite = "markitwrite.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
